Removed auto-placeholder logic completely
After hours of playing around on paper, I came to the conclusion that making placeholders work the way I wanted them to bug free is just too difficult. Predicting the area to leave space is magnitudes more complicated than the initial attempt of leaving something like EMPTY_CUT size. Because of the 2D nature of graphs, figuring out a calculation that determines where a rightmost edge might be is ridiculously more complicated than a static value.

I think a more productive approach will be to develop the placeholder approach (i.e. an invisible object that can't be clicked on and moved, but helps to define the boundaries of a cut).

For now, debug key 4 adds a letter like the rest, only its a space symbol, making it essentially invisible unless hovered over. I'm planning on making the placeholders not even hoverable nor collideable, and perhaps implementing a mouse draggable resizing feature around cut edges, which manipulates these placeholders as needed. It shouldn't be too difficult to accomplish, but it's low priority for now.