Altered selection mechanics a bit
Not sure if I like this change, but if you have an existing selection now, and you try to move a node that isn't selected, the selection is cleared and you only operate on a single node. Previously, the node would try to be added to the selection, which I wasn't too happy with either. The previous way had an issue with the selection growing and lingering too long; it's hard to explain, but if you had any selection and moved multiple siblings one by one, the selection would grow and grow and that forces the user to clear it manually.