Second bloom function for adding
Adding a node will use the same bloom trick as moving to appear a bit more intelligent than it actually is. This bloom will be checked against to see if a node can be added in the general vicinity of the mouse to a place where no collision or parental growth is needed. Obviously, this won't always happen, but the idea with this bloom is to make adding a statement to an empty cut completely painless: i.e. the mouse can be anywhere inside it and it will still hopefully find the central spot instead of needing to move the parent around.