Buggy attempt at changing parent
Hierarchy seems to be okay. Learned that the invisible root node should
not be set as a parentItem through Qt -- this will make the child node
invisible as well. The workaround is to setParentItem(0), which makes it
top level in the case where it's supposed to be on the page.

Actually moving stuff around is really buggy. The offset from updating
the drawBox to the new coordinate system is really incorrect, and it's
late in the night so I can't figure out why.

Further, there should be some sort of snapping prediction like in the
add function, so dropped nodes can snap to reasonable places. Also in
the realm of aesthetics: should probably make the blue border (currently
only viewable when the debug bounds are turned on) that signifies the
target parent visible to the end user.