Added a flag to remove highlight on mouse move
Might be more intuitive, since you don't have to reposition the mouse on a highlighted node and then remove it again just to dehighlight it if it was highlighted by keyboard actions (e.g. removeSurroundingCut will force highlight the parent after wards to make double cut removals easy to do) even though the mouse is not necessarily on top of that node.

This new behavior is likely introducing a lot of bugs and needs updating.